+++ fpdf/util/syntax.py
@@ -1,8 +1,8 @@
 """PDF Syntax Helpers
 
 Functions in this module take variable input and produce PDF Syntax features
-as they are described in the Adobe PDF Reference Manual, found
-[here](http://www.adobe.com/content/dam/Adobe/en/devnet/acrobat/pdfs/pdf_reference_1-7.pdf).
+as they are described in the Adobe PDF Reference Manual, found here:
+http://www.adobe.com/content/dam/Adobe/en/devnet/acrobat/pdfs/pdf_reference_1-7.pdf
 
 Most of what happens in a PDF happens in objects, which are formatted like so:
 <pre>
@@ -71,7 +71,7 @@ remains something to be looked into.
 def create_name(name):
     if name.startswith("/"):
         name = name[1:]
-    return "".join(["/", name[0].upper(), name[1:]])
+    return f"/{name[0].upper()}{name[1:]}"
 
 
 def clear_empty_fields(d):
@@ -101,7 +101,7 @@ def create_dictionary_string(
     return "".join(
         [
             open_dict,
-            field_join.join([key_value_join.join(f) for f in dict_.items()]),
+            field_join.join(key_value_join.join(f) for f in dict_.items()),
             close_dict,
         ]
     )