+++ CHANGELOG.md
@@ -7,13 +7,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/),
 and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 
-## [2.2.1] - Not released yet
+## [2.3.1] - not released yet
 ### Added
+- `FPDF.set_margin` : new method to set the document right, left, top & bottom margins to the same value
+
+## [2.3.0] - 2021-01-29
+Many thanks to [@eumiro](https://github.com/PyFPDF/fpdf2/pulls?q=is%3Apr+author%3Aeumiro) & [@fbernhart](https://github.com/PyFPDF/fpdf2/pulls?q=is%3Apr+author%3Aeumiro) for their contributions to make `fpdf2` code cleaner!
+### Added
+- `FPDF.unbreakable` : a new method providing a context-manager in which automatic page breaks are disabled.
+  _cf._ https://pyfpdf.github.io/fpdf2/PageBreaks.html
+- `FPDF.epw` & `FPDF.eph` : new `@property` methods to retrieve the **effective page width / height**, that is the page width / height minus its horizontal / vertical margins.
 - `FPDF.image` now accepts also a `Pillow.Image.Image` as input
 - `FPDF.multi_cell` parameters evolve in order to generate tables with multiline text in cells:
   * its `ln` parameter now accepts a value of `3` that sets the new position to the right without altering vertical offset
   * a new optional `max_line_height` parameter sets a maximum height of each sub-cell generated
-- documentation on how to add content to existing PDFs
+- new documentation pages : how to add content to existing PDFs, HTML, links, tables, text styling & page breaks
+- all PDF samples are now validated using 3 different PDF checkers
 ### Fixed
 - `FPDF.alias_nb_pages`: fixed this feature that was broken since v2.0.6
 - `FPDF.set_font`: fixed a bug where calling it several times, with & without the same parameters,